Update createTable.sql
The id field has to be larger than an INT as Twitter's account ids can now be bigger than the maximum value. Thanks to Chris Millard (chrismillard.com) for helping with this!

+++ createTable.sql
@@ -1,5 +1,5 @@
 CREATE TABLE IF NOT EXISTS `followers` (
-  `id` int(11) NOT NULL,
+  `id` bigint(20) NOT NULL,
   `username` text NOT NULL,
   `screen_name` text NOT NULL,
   PRIMARY KEY  (`id`)
